Support IBM AIX
Add support for IBM AIX:
  - Allow overriding Makefile variables
  - Respect LDFLAGS in Makefile
  - Check if obj directory exists before attempting removal
  - Add portable implementation of asprintf
  - Rename function int16 (as it may be an OS defined type)

Tested with IBM XL C/C++ V16.1.0.19, IBM Open XL C/C++ V17.1.2.11,
and GNU GCC 12.3.0 on IBM AIX 7.2 and AIX 7.3 with GNU Make 4.4.1:

env CC="/opt/IBM/xlC/16.1.0/bin/xlc" \
  CFLAGS="-q64 -O3" LDFLAGS="-Wl,-b64" gmake

env CC="/opt/IBM/openxlC/17.1.2/bin/ibm-clang" \
  CFLAGS="-m64 -O3 -flto" LDFLAGS="-Wl,-b64" gmake

env CC="gcc-12" CFLAGS="-maix64 -O3" LDFLAGS="-Wl,-b64" gmake

+++ Makefile
@@ -1,10 +1,11 @@
-CC=cc
-CFLAGS=-O3 -flto -Wall -g -Werror=implicit-function-declaration -Werror=int-conversion
-LDLIBS=-lm
-INSTALL=install
-PREFIX=/usr
+CC?=cc
+CFLAGS?=-O3 -flto -Wall -g -Werror=implicit-function-declaration -Werror=int-conversion
+LDLIBS?=-lm
+INSTALL?=install
+PREFIX?=/usr
 
 OBJS=\
+ asprintf.o\
  cpu.o\
  loader.o\
  main.o\
@@ -22,7 +23,7 @@ OBJS=\
 all: obj emu2
 
 emu2: $(OBJS:%=obj/%)
-	$(CC) $(CFLAGS) -o $@ $^ $(LDLIBS)
+	$(CC) $(CFLAGS) -o $@ $^ $(LDFLAGS) $(LDLIBS)
 
 obj/%.o: src/%.c
 	$(CC) $(CFLAGS) -c -o $@ $<
@@ -32,7 +33,7 @@ obj:
 clean:
 	rm -f $(OBJS:%=obj/%)
 	rm -f emu2
-	rmdir obj
+	test -d obj && rmdir obj || true
 
 install: emu2
 	$(INSTALL) -d $(DESTDIR)${PREFIX}/bin
